storage: move compression to its own file
