+++ src/storage/mod.rs
@@ -1,6 +1,8 @@
+mod compression;
 mod database;
 pub(crate) mod s3;
 
+pub use self::compression::{compress, decompress, CompressionAlgorithm, CompressionAlgorithms};
 pub(crate) use self::database::DatabaseBackend;
 pub(crate) use self::s3::S3Backend;
 use crate::db::Pool;
@@ -11,67 +13,10 @@ use std::{
     collections::{HashMap, HashSet},
     ffi::OsStr,
     fmt, fs,
-    io::Read,
     path::{Path, PathBuf},
 };
 
 const MAX_CONCURRENT_UPLOADS: usize = 1000;
-
-pub type CompressionAlgorithms = HashSet<CompressionAlgorithm>;
-
-macro_rules! enum_id {
-    ($vis:vis enum $name:ident { $($variant:ident = $discriminant:expr,)* }) => {
-        #[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
-        $vis enum $name {
-            $($variant = $discriminant,)*
-        }
-
-        impl $name {
-            #[cfg(test)]
-            const AVAILABLE: &'static [Self] = &[$(Self::$variant,)*];
-        }
-
-        impl fmt::Display for CompressionAlgorithm {
-            fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-                match self {
-                    $(Self::$variant => write!(f, stringify!($variant)),)*
-                }
-            }
-        }
-
-        impl std::str::FromStr for CompressionAlgorithm {
-            type Err = ();
-            fn from_str(s: &str) -> Result<Self, Self::Err> {
-                match s {
-                    $(stringify!($variant) => Ok(Self::$variant),)*
-                    _ => Err(()),
-                }
-            }
-        }
-
-        impl std::convert::TryFrom<i32> for CompressionAlgorithm {
-            type Error = i32;
-            fn try_from(i: i32) -> Result<Self, Self::Error> {
-                match i {
-                    $($discriminant => Ok(Self::$variant),)*
-                    _ => Err(i),
-                }
-            }
-        }
-    }
-}
-
-enum_id! {
-    pub enum CompressionAlgorithm {
-        Zstd = 0,
-    }
-}
-
-impl Default for CompressionAlgorithm {
-    fn default() -> Self {
-        CompressionAlgorithm::Zstd
-    }
-}
 
 #[derive(Clone, Debug, PartialEq, Eq, Hash)]
 pub(crate) struct Blob {
@@ -231,28 +176,6 @@ trait StorageTransaction {
     fn complete(self: Box<Self>) -> Result<(), Error>;
 }
 
-// public for benchmarking
-pub fn compress(content: impl Read, algorithm: CompressionAlgorithm) -> Result<Vec<u8>, Error> {
-    match algorithm {
-        CompressionAlgorithm::Zstd => Ok(zstd::encode_all(content, 9)?),
-    }
-}
-
-pub fn decompress(
-    content: impl Read,
-    algorithm: CompressionAlgorithm,
-    max_size: usize,
-) -> Result<Vec<u8>, Error> {
-    // The sized buffer prevents a malicious file from decompressing to multiple times its size.
-    let mut buffer = crate::utils::sized_buffer::SizedBuffer::new(max_size);
-
-    match algorithm {
-        CompressionAlgorithm::Zstd => zstd::stream::copy_decode(content, &mut buffer)?,
-    }
-
-    Ok(buffer.into_inner())
-}
-
 fn detect_mime(file_path: &Path) -> Result<&'static str, Error> {
     let mime = mime_guess::from_path(file_path)
         .first_raw()
@@ -399,65 +322,6 @@ mod test {
     }
 
     #[test]
-    fn test_compression() {
-        let orig = "fn main() {}";
-        for alg in CompressionAlgorithm::AVAILABLE {
-            println!("testing algorithm {}", alg);
-
-            let data = compress(orig.as_bytes(), *alg).unwrap();
-            let blob = Blob {
-                mime: "text/rust".into(),
-                content: data.clone(),
-                path: "main.rs".into(),
-                date_updated: Utc::now(),
-                compression: Some(*alg),
-            };
-            test_roundtrip(std::slice::from_ref(&blob));
-            assert_eq!(
-                decompress(data.as_slice(), *alg, std::usize::MAX).unwrap(),
-                orig.as_bytes()
-            );
-        }
-    }
-
-    #[test]
-    fn test_decompression_too_big() {
-        const MAX_SIZE: usize = 1024;
-
-        let small = &[b'A'; MAX_SIZE / 2] as &[u8];
-        let exact = &[b'A'; MAX_SIZE] as &[u8];
-        let big = &[b'A'; MAX_SIZE * 2] as &[u8];
-
-        for &alg in CompressionAlgorithm::AVAILABLE {
-            let compressed_small = compress(small, alg).unwrap();
-            let compressed_exact = compress(exact, alg).unwrap();
-            let compressed_big = compress(big, alg).unwrap();
-
-            // Ensure decompressing within the limit works.
-            assert_eq!(
-                small.len(),
-                decompress(compressed_small.as_slice(), alg, MAX_SIZE)
-                    .unwrap()
-                    .len()
-            );
-            assert_eq!(
-                exact.len(),
-                decompress(compressed_exact.as_slice(), alg, MAX_SIZE)
-                    .unwrap()
-                    .len()
-            );
-
-            // Ensure decompressing a file over the limit returns a SizeLimitReached error.
-            let err = decompress(compressed_big.as_slice(), alg, MAX_SIZE).unwrap_err();
-            assert!(err
-                .downcast_ref::<std::io::Error>()
-                .and_then(|io| io.get_ref())
-                .and_then(|err| err.downcast_ref::<crate::error::SizeLimitReached>())
-                .is_some());
-        }
-    }
-
-    #[test]
     fn test_mime_types() {
         check_mime(".gitignore", "text/plain");
         check_mime("hello.toml", "text/toml");
@@ -476,18 +340,5 @@ mod test {
         let detected_mime = detect_mime(Path::new(&path));
         let detected_mime = detected_mime.expect("no mime was given");
         assert_eq!(detected_mime, expected_mime);
-    }
-
-    #[test]
-    fn test_compression_try_from_is_exhaustive() {
-        use std::convert::TryFrom;
-
-        let a = CompressionAlgorithm::Zstd;
-        match a {
-            CompressionAlgorithm::Zstd => {
-                assert_eq!(a, CompressionAlgorithm::try_from(a as i32).unwrap());
-                assert_eq!(a, a.to_string().parse().unwrap());
-            }
-        }
     }
 }
